add gatsby-node.ts
Current plugin `gatsby-plugin-codegen` can't generate types from graphql
inside `gatsby-node.js`. Consider change to `gatsby-plugin-typegen`,
because they have plan to support `gatsby-node.js` graphql types in the
future.
