+++ gatsby-node.js
@@ -2,177 +2,7 @@
 // Read the LICENSE file in the repository root for full license text.
 
 /* eslint-disable @typescript-eslint/no-var-requires */
-const path = require(`path`);
-const { createFilePath } = require(`gatsby-source-filesystem`);
-const kebabCase = require("lodash.kebabcase");
+require("source-map-support").install();
+require("ts-node").register({ files: true, transpileOnly: true });
 
-exports.createPages = async ({ graphql, actions, reporter }) => {
-  const { createPage } = actions;
-
-  const PostTemplate = path.resolve(`./src/templates/Post.tsx`);
-  const BlogTemplate = path.resolve(`./src/templates/Blog.tsx`);
-
-  const result = await graphql(`
-    query GatsbyNode {
-      allPosts: allMarkdownRemark(
-        limit: 1000
-        sort: { fields: frontmatter___date, order: DESC }
-      ) {
-        nodes {
-          id
-          fields {
-            slug
-          }
-        }
-      }
-
-      allCategories: allMarkdownRemark {
-        group(field: frontmatter___categories) {
-          ...GroupInfo
-        }
-      }
-
-      allTags: allMarkdownRemark {
-        group(field: frontmatter___tags) {
-          ...GroupInfo
-        }
-      }
-
-      allLang: allMarkdownRemark {
-        group(field: frontmatter___lang) {
-          ...GroupInfo
-        }
-      }
-    }
-
-    fragment GroupInfo on MarkdownRemarkGroupConnection {
-      fieldValue
-      totalCount
-    }
-  `);
-
-  if (result.errors) {
-    reporter.panicOnBuild(
-      `There was an error when loading blog posts`,
-      result.errors
-    );
-    return;
-  }
-
-  const posts = result.data.allPosts.nodes;
-  const categories = result.data.allCategories.group;
-  const tags = result.data.allTags.group;
-  const langs = result.data.allLang.group;
-
-  if (posts.length <= 0) {
-    reporter.warn(`There is no posts!`);
-    return;
-  }
-
-  posts.forEach((post, index) => {
-    const newerId = index === 0 ? null : posts[index - 1].id;
-    const olderId = index === posts.length - 1 ? null : posts[index + 1].id;
-
-    const path = post.fields.slug;
-    reporter.info(`Creating page ${path}, type Post`);
-
-    createPage({
-      path,
-      component: PostTemplate,
-      context: {
-        id: post.id,
-        newerId,
-        olderId
-      }
-    });
-  });
-
-  const postPerPage = 5;
-
-  const createUpdatePage = ({ slug, type, postCount, filterValue }) => {
-    const numPages = Math.ceil(postCount / postPerPage);
-
-    const paths = [];
-    if (slug !== undefined) {
-      paths.push(slug);
-    }
-
-    if (filterValue !== undefined) {
-      paths.push(kebabCase(filterValue));
-    }
-
-    let path = "/";
-    if (paths.length > 0) {
-      path += `${paths.join("/")}/`;
-    }
-
-    Array.from({ length: numPages }).forEach((_, index) => {
-      const templatePath = path;
-      if (index > 0) {
-        path += `${index + 1}`;
-      }
-
-      reporter.info(`Creating page ${path}, type ${type}`);
-
-      createPage({
-        path,
-        component: BlogTemplate,
-        context: {
-          limit: postPerPage,
-          skip: index * postPerPage,
-          page: index + 1,
-          templatePath,
-          numPages,
-          type,
-          filterValue
-        }
-      });
-    });
-  };
-
-  createUpdatePage({
-    type: "Index",
-    postCount: posts.length
-  });
-
-  categories.forEach(category => {
-    createUpdatePage({
-      slug: "category",
-      type: "Category",
-      postCount: category.totalCount,
-      filterValue: category.fieldValue
-    });
-  });
-
-  tags.forEach(tag => {
-    createUpdatePage({
-      slug: "tag",
-      type: "Tag",
-      postCount: tag.totalCount,
-      filterValue: tag.fieldValue
-    });
-  });
-
-  langs.forEach(lang => {
-    createUpdatePage({
-      slug: "lang",
-      type: "Language",
-      postCount: lang.totalCount,
-      filterValue: lang.fieldValue
-    });
-  });
-};
-
-exports.onCreateNode = ({ node, actions, getNode }) => {
-  const { createNodeField } = actions;
-
-  if (node.internal.type === `MarkdownRemark`) {
-    const value = createFilePath({ node, getNode });
-
-    createNodeField({
-      name: `slug`,
-      node,
-      value
-    });
-  }
-};
+module.exports = require("./gatsby-node.ts");
